requirements.txt
python-gitlab==3.15.0
urllib3==2.0.7
PyGithub==1.59.1
rich==13.8.0
ell-ai[all]
